convert games to app dir
