+++ apps/games/app/api/hello.ts
