+++ apps/games/app/preload/page.tsx
@@ -1,4 +1,7 @@
+'use client'
+
 import type { NextPage } from 'next'
+
 import Head from 'next/head'
 import { useEffect, useState } from 'react'
