+++ apps/games/package.json
@@ -10,13 +10,13 @@
 	},
 	"dependencies": {
 		"@faker-js/faker": "7.4.0",
-		"@radix-ui/react-icons": "1.1.1",
 		"framer-motion": "6.2.3",
 		"just-shuffle": "4.1.1",
 		"lib": "workspace:*",
 		"next": "13.3.4",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
+		"react-icons": "4.7.1",
 		"ui": "workspace:*"
 	},
 	"devDependencies": {
