--- apps/games/pages/_app.tsx
@@ -1,14 +1,0 @@
-import Layout from '@/components/Layout'
-import type { AppProps } from 'next/app'
-import 'ui/globals.css'
-import '../styles/globals.css'
-
-function MyApp({ Component, pageProps }: AppProps) {
-	return (
-		<Layout>
-			<Component {...pageProps} />
-		</Layout>
-	)
-}
-
-export default MyApp
